### 2024-10-22 #7356 @fabian-s
Sorry, have to decline. Don't know JAX and don't know much optimization.
