### 2024-11-07 62d8db1
Add thin and progress kwargs to fitters
