### 2024-11-09 #7356 @crvernon
:wave: @jbytecode can you take this one on as editor?
