### 2024-11-10 #7356 @jbytecode
@crvernon - sure, gladly. But I have a bunch of issues about the submission.

Do you think a scope check voting needs to be conducted for this study?

Because,

- The commit history goes back to Oct 3rd, 2014. I think it's a young project.
- The README.md has not enough information, so the documentation may be improved before considering any further review tasks.

These issues might also likely require subsequently scholar effort voting.

What do you think?
