### 2024-11-10 #7356 @AlanPearl
@jbytecode
- The young age is due to it being a resubmission. Initially, the subpackages were their own repositories, which are still hosted on my GitHub with their own histories that date back roughly a year (at least for multigrad which is the oldest of the three subpackages).
- Sorry that the README is not sufficient. My typical philosophy is having a very short one containing an easy-to-find link to the readthedocs page, which contains the extensive documentation. Let me know if there is anything specific that needs to go in the README.
