### 2024-11-10 #7356 @jbytecode
@AlanPearl - Thank you for the quick response. It seems the umbrella package wraps other packages. Are they yours? If so it's a better approach to attach the package names in the README.md file as dependencies and let our reviewers to consider them.

And it's time to hear from our track editor. Thank you in advance.
