+++ README.md
@@ -6,3 +6,13 @@ Parallelization and optimization of differentiable and many-parameter models
 
 ## Documentation
 Online documentation is available at [diffopt.readthedocs.io](https://diffopt.readthedocs.io/en/latest).
+
+## JOSS Reviewer Information
+
+### Finding the full git repo histories
+- `diffopt` combined the scope of several subpackages, `multigrad`, `kdescent`, and `multiswarm`. This was intended to satisfy [initial concerns that an individual submission did not satisfy the scholarly effort requirement](https://github.com/openjournals/joss-reviews/issues/7218).
+- I developed these subpackages separately before combining them into a single repository, which is why `diffopt` has such a short history.
+- Their full respective histories can still be found at:
+  - https://github.com/AlanPearl/multigrad
+  - https://github.com/AlanPearl/kdescent
+  - https://github.com/AlanPearl/multiswarm
