### 2024-11-15 #7356 @crvernon
@jbytecode FYI...here is the previous submission that was withdrawn so that the current submission could be reviewed as a complete product:

https://github.com/openjournals/joss-reviews/issues/7218
