### 2024-11-15 #7356 @jbytecode
@crvernon - Now, that makes sense.

@AlanPearl  - Could you please provide a list of the dependencies of the submitted package in README.md file so our reviewers can take them into consideration in the review process.
