### 2024-11-15 #7356 @AlanPearl
@jbytecode I updated the README.md with the dependency info on the paper branch. Does that look good?
