### 2024-11-15 #7356 @jbytecode
@AlanPearl - Do you have any suggestions for the potential reviewers? Editorial bot shows us many similar submissions and I think the author of the submission

https://github.com/openjournals/joss-reviews/issues/3525

can be invited as a reviewer. Would you provide a list of potential reviewers for your submission? You can also use the [list of people](https://reviewers.joss.theoj.org/lookup) to search over people who volunteer to review. Please mention their GitHub handles without using the `@` character (because we don't want to bother them by being too much verbose).

Thank you in advance.
