### 2024-11-19 #7356 @AlanPearl
@jbytecode I just scanned through editorialbot suggestions, and I agree that @(landreman)'s submission is the closest. Next would probably be @(ewu63), submitting author of #2564. Their method doesn't use jax/autodiff, but it could still be useful to have someone working on parallel gradient-based optimization problems from another angle take a look.

Other possible suggestions of people with similar interests in optimization + astrophysics I found on the JOSS list of people:
- @(htjb)
- @(astrogilda)
- @(athulpg007)
- @(DingXuefeng)
