### 2024-11-20 #7356 @jbytecode
👋👋👋 Dear @htjb 👋👋👋

**[DiffOpt: Parallel optimization of Jax models](https://github.com/openjournals/joss-reviews/issues/7356)**

You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/7356).

This is the pre-review issue. After setting at least 2 reviewers we will start the review process in a separate thread. In that thread, there will be around 20 check items for each single reviewer.

Thank you in advance!
