### 2024-11-23 #7356 @ewu63
> 👋👋👋 Dear @ewu63 👋👋👋
>
> Would you be willing to assist in reviewing this submission for **JOSS** (**Journal of Open Source Software**)?
>
> JOSS publishes articles about open source research software. The submission I'd like you to review is titled:
>
> **[DiffOpt: Parallel optimization of Jax models](https://github.com/openjournals/joss-reviews/issues/7356)**
>
> You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/7356).
>
> The review process at JOSS is unique: it takes place in a GitHub issue, is open, and author-reviewer-editor conversations are encouraged. If you have any questions please let me know.
>
> This is the pre-review issue. After setting at least 2 reviewers we will start the review process in a separate thread. In that thread, there will be around 20 check items for each single reviewer.
>
>
> Thank you in advance!
>
>
>

I am also happy to be a reviewer.
