### 2024-11-24 #7522 @editorialbot
[REVIEW]: DiffOpt: Parallel optimization of Jax models
**Submitting author:** @AlanPearl (<a href="http://orcid.org/0000-0001-9820-9619">Alan Pearl</a>)
**Repository:** https://github.com/AlanPearl/diffopt
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.1.1
