### 2024-11-25 AlanPearl/diffopt#1 @landreman
References in JOSS paper
For the [JOSS review](https://github.com/openjournals/joss-reviews/issues/7522):

Googling around quickly, it looks like another project to connect mpi with jax is https://github.com/mpi4jax/mpi4jax. How about citing their [paper](https://joss.theoj.org/papers/10.21105/joss.03419) and adding a few words in your manuscript about the similarities/differences?

Also some info is missing on the Scott reference.
