### 2024-11-25 AlanPearl/diffopt#2 @landreman
contributing
For the [JOSS review](https://github.com/openjournals/joss-reviews/issues/7522)- The journal asks for the following:

"Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support"
