### 2024-11-25 AlanPearl/diffopt#3 @landreman
Tests
For the [JOSS review](https://github.com/openjournals/joss-reviews/issues/7522)- The journal requires the following:

"Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?"

I didn't see discussion in the docs of how to run the tests. Consider adding a Github Actions workflow to run the tests automatically?
