### 2024-11-25 1b87a3d
Try conda install mpi4py in workflow
