### 2024-11-26 94694dc
Skip unit tests requiring mpi4py if not installed
