### 2024-11-26 dc2c043
Add manual testing instructions to README
