### 2024-11-26 43d4b57
Create test_package workflow (#4)
* Create test_package workflow

* Skip unit tests requiring mpi4py if not installed

* Add manual testing instructions to README
