### 2024-11-26 AlanPearl/diffopt#3 @AlanPearl
Good catch. I just now pushed an automatic testing workflow (plus manual testing instructions) to the main branch
