### 2024-11-26 7e74091
Create CONTRIBUTING.md and move testing instructions here
