### 2024-11-26 805c0fa
Link to contributing guidelines in docs
