+++ docs/requirements.txt
@@ -1,5 +1,6 @@
 sphinx_rtd_theme
 nbsphinx
+myst_parser
 IPython
 matplotlib
 seaborn
