### 2024-11-26 f9e0146
Get better BibTeX entry for Scott:1992
