+++ paper/paper.bib
@@ -62,11 +62,16 @@
 }
 
 @book{Scott:1992,
-       author = {{Scott}, D.~W.},
-        title = "{Multivariate Density Estimation}",
-         year = 1992,
-       adsurl = {https://ui.adsabs.harvard.edu/abs/1992mde..book.....S},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+ title={Multivariate Density Estimation: Theory, Practice, and Visualization},
+ author={Scott, D.W.},
+ isbn={9780471547709},
+ lccn={lc91043950},
+ series={A Wiley-interscience publication},
+ url={https://books.google.com/books?id=7crCUS_F2ocC},
+ year={1992},
+ publisher={Wiley},
+ adsurl={https://ui.adsabs.harvard.edu/abs/1992mde..book.....S},
+ adsnote={Provided by the SAO/NASA Astrophysics Data System}
 }
 
 @book{Cramer:1954,
