### 2024-11-26 AlanPearl/diffopt#1 @AlanPearl
- The Scott reference is now fixed: [f9e0146](https://github.com/AlanPearl/diffopt/commit/f9e0146115e432104a50080ea178adef363e46c6)
- Thanks for bringing up mpi4jax, I will definitely cite it and add some clarification on why to use one or the other...
