+++ paper/paper.bib
@@ -95,3 +95,54 @@
   keywords={Particle swarm optimization;Birds;Educational institutions;Marine animals;Testing;Humans;Genetic algorithms;Optimization methods;Artificial neural networks;Performance evaluation},
   doi={10.1109/ICNN.1995.488968}
 }
+
+@article{Hafner:2021,
+ doi = {10.21105/joss.03419},
+ url = {https://doi.org/10.21105/joss.03419},
+ year = {2021},
+ publisher = {The Open Journal},
+ volume = {6},
+ number = {65},
+ pages = {3419},
+ author = {Dion Häfner and Filippo Vicentini},
+ title = {mpi4jax: Zero-copy MPI communication of JAX arrays},
+ journal = {Journal of Open Source Software}
+}
+
+@article{Gray:2019,
+    Author = {Justin S. Gray and John T. Hwang
+              and Joaquim R. R. A. Martins
+              and Kenneth T. Moore and Bret A. Naylor},
+    Doi = {10.1007/s00158-019-02211-z},
+    Journal = {Structural and Multidisciplinary Optimization},
+    Month = {April},
+    Number = {4},
+    Pages = {1075--1104},
+    Title = {{OpenMDAO}: An open-source framework
+             for multidisciplinary design, analysis,
+             and optimization},
+    Volume = {59},
+    Year = {2019}
+}
+
+@INPROCEEDINGS{Li:2005,
+  author={Bo Li and Koichi Wada},
+  booktitle={PACRIM. 2005 IEEE Pacific Rim Conference on Communications, Computers and signal Processing, 2005.},
+  title={Parallelizing particle swarm optimization},
+  year={2005},
+  volume={},
+  number={},
+  pages={288-291},
+  keywords={Particle swarm optimization;Large-scale systems;Artificial neural networks;Clustering algorithms;Particle tracking;Computer science;Algorithm design and analysis;Optimization methods;Convergence;Pattern recognition},
+  doi={10.1109/PACRIM.2005.1517282}
+}
+
+@ARTICLE{Blank:2020,
+    author={J. {Blank} and K. {Deb}},
+    journal={IEEE Access},
+    title={pymoo: Multi-Objective Optimization in Python},
+    year={2020},
+    volume={8},
+    number={},
+    pages={89497-89509},
+}
