### 2024-12-03 AlanPearl/diffopt#1 @AlanPearl
I have pushed changes to the statement of need in a2b6e83 including references to similar past efforts, including the two mentioned in this thread. How does this look?
