### 2024-12-04 AlanPearl/diffopt#1 @landreman
Looks good to me
