### 2024-12-04 #7522 @landreman
All the items in the checklist are complete so I approve the paper for publication. Is there anything more I need to do to sign off?
