### 2024-12-05 AlanPearl/diffopt#1 @ewu63
Good with me as well
