### 2024-12-05 AlanPearl/diffopt#1 @jbytecode
Can we please close the resolved issues?
