### 2024-12-05 AlanPearl/diffopt#1 @AlanPearl
Yes, this is now closed
