### 2024-12-06 #7522 @ewu63
@jbytecode the only comment I had was [here](https://github.com/AlanPearl/diffopt/issues/1#issuecomment-2508836171) regarding references. The main module of the library, `multigrad`, propagates jax-computed Jacobians under MPI, and is used by `kdescent` to perform stochastic gradient descent. `multiswarm` on the other hand implements a parallel PSO algorithm, which is neither novel in theory nor in implementation. However, while similar capabilities can be obtained via established libraries such as mpi4jax, this library does offer some conveniences in providing wrappers for specific applications such as running `adam`, and has been used to produce published results in astrophysics.
I have locally installed the package and verified that the examples/tests run and achieve the speedups claimed in the paper, and the tutorials/documentation are well written. The paper itself is also well written and I recommend this paper for publication as is.
