### 2024-12-06 1c86f15
Merge software changes into paper branch (#6)
* Add thin and progress kwargs to fitters

* Create test_package workflow (#4)

* Create test_package workflow

* Skip unit tests requiring mpi4py if not installed

* Add manual testing instructions to README

* Create CONTRIBUTING.md and move testing instructions here

* Link to contributing guidelines in docs
