+++ docs/source/conf.py
@@ -25,6 +25,7 @@ extensions = [
     "sphinx.ext.napoleon",
     "sphinx.ext.viewcode",
     "nbsphinx",
+    "myst_parser",
 ]
 
 templates_path = ['_templates']
