### 2024-12-07 #7522 @AlanPearl
Done. Here is the requested info:
- Software release: [v1.1.1](https://github.com/AlanPearl/diffopt/releases/tag/v1.1.1)
- [10.5281/zenodo.14291795](https://zenodo.org/records/14291795)
