### 2024-12-08 #7522 @jbytecode
@AlanPearl - The metadata of manuscript should match those in the Zenodo archive, please add contributors to the authors list with the proper information including their DOIs. Thank you in advance.
