+++ paper/paper.bib
@@ -71,7 +71,8 @@
  year={1992},
  publisher={Wiley},
  adsurl={https://ui.adsabs.harvard.edu/abs/1992mde..book.....S},
- adsnote={Provided by the SAO/NASA Astrophysics Data System}
+ adsnote={Provided by the SAO/NASA Astrophysics Data System},
+ doi={10.2307/1270280}
 }
 
 @book{Cramer:1954,
@@ -145,4 +146,5 @@
     volume={8},
     number={},
     pages={89497-89509},
+    doi={10.1109/access.2020.2990567}
 }
