### 2024-12-09 #7522 @jbytecode
@AlanPearl - It looks like it's resolved, thank you. Please let me have a full read of the manuscript and check my editorial tasks. This may take a couple of days.
