### 2024-12-13 #7522 @jbytecode
@AlanPearl - The manuscript looks almost perfect. I've just sent a pull request that includes minor fixes in both the manuscript and the bibliography (https://github.com/AlanPearl/diffopt/pull/7). Please review the PR, if you're okay with them, please merge.
Thank you in advance.
