### 2024-12-15 #7522 @jbytecode
@AlanPearl - The manuscript looks good to me. Please perform a full read of the paper and fix any thing if you think it's wrong. Please tell me when you've done with them. Thank you in advance.
