### 2024-12-17 #7522 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
