### 2024-12-20 #7522 @crvernon
Very clean paper.  Thank you!
